Implement creating a texture from an image structure
Add make_sdl_texture_from_image which creates an SDL_Texture from
an image structure. It's a bit gross having to pass in a renderer
but I don't see an easy way to fix this.

The colourdepth is also hardcoded, which is probably wrong.

The masks will only work on certain architectures, so this will need
to be fixed as well.